backend/requirements.txt
# --- START OF FILE backend/requirements.txt ---
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
requests==2.32.3
Markdown==3.6
psycopg2-binary==2.9.9
# --- END OF FILE backend/requirements.txt ---